# serviflex-server/docs/swagger.yaml
definitions:
  controllers.ImagemInput:
    properties:
      imagem_url:
        type: string
    type: object
  models.Agendamento:
    properties:
      cliente_id:
        type: string
      data_hora:
        type: string
      id:
        type: string
      procedimento:
        type: string
      profissional_id:
        type: string
    type: object
...
  models.Horario:
    properties:
      dia_semana:
        type: string
      disponivel:
        type: boolean
      hora_fim:
        type: string
      hora_inicio:
        type: string
      id:
        type: string
      profissional_id:
        type: string
    type: object
  models.HorarioInput:
    properties:
      dias_semana:
        items:
          type: string
        type: array
      hora_fim:
        type: string
      hora_inicio:
        type: string
      profissional_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /agendamentos/profissional/{id}:
    get:
      parameters:
      - description: ID do profissional
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Agendamento'
            type: array
      summary: Listar agendamentos por profissional
      tags:
      - Agendamentos
...
  /horarios:
    post:
      consumes:
      - application/json
      parameters:
      - description: Horário de atendimento
        in: body
        name: horario
        required: true
        schema:
          $ref: '#/definitions/models.HorarioInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar horários
      tags:
      - Horários
  /horarios/{id}:
    get:
      parameters:
      - description: ID do profissional
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Horario'
            type: array
      summary: Listar horários por profissional
      tags:
      - Horários
...
  /upload/{tipo}/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Tipo do recurso (profissional ou procedimento)
        in: path
        name: tipo
        required: true
        type: string
      - description: ID do recurso
        in: path
        name: id
        required: true
        type: string
      - description: URL da imagem
        in: body
        name: imagem
        required: true
        schema:
          $ref: '#/definitions/controllers.ImagemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar imagem do recurso
      tags:
      - Upload
swagger: "2.0"